requirements.txt
discord.py
pyyaml
requests
psutil
pylint
pymongo[srv]
